refactor(tests): split zoo auth tests to ai-zoo-auth.test.ts (200-line rule)
ai-routes.test.ts exceeded the 200-line ESLint limit after adding
security tests. Moved the #483 zoo-auth test suite to its own file.
